[refactor] Test renamed to TestEntity
